* Refactoring locales
