+++ muikku-timed-notifications-plugin/src/main/java/fi/otavanopisto/muikku/plugins/timed/notifications/strategies/NeverLoggedInNotificationStrategy.java
@@ -3,7 +3,6 @@ package fi.otavanopisto.muikku.plugins.timed.notifications.strategies;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
 import java.util.Locale;
@@ -24,20 +23,15 @@ import java.time.OffsetDateTime;
 import fi.otavanopisto.muikku.controller.PluginSettingsController;
 import fi.otavanopisto.muikku.i18n.LocaleController;
 import fi.otavanopisto.muikku.model.users.UserEntity;
-import fi.otavanopisto.muikku.model.users.UserGroupEntity;
 import fi.otavanopisto.muikku.plugins.activitylog.ActivityLogController;
 import fi.otavanopisto.muikku.plugins.activitylog.model.ActivityLogType;
 import fi.otavanopisto.muikku.plugins.timed.notifications.NeverLoggedInNotificationController;
 import fi.otavanopisto.muikku.plugins.timed.notifications.NotificationController;
 import fi.otavanopisto.muikku.schooldata.SchoolDataIdentifier;
-import fi.otavanopisto.muikku.schooldata.entity.GroupUser;
 import fi.otavanopisto.muikku.schooldata.entity.User;
-import fi.otavanopisto.muikku.schooldata.entity.UserGroup;
 import fi.otavanopisto.muikku.search.SearchResult;
 import fi.otavanopisto.muikku.users.UserController;
 import fi.otavanopisto.muikku.users.UserEntityController;
-import fi.otavanopisto.muikku.users.UserGroupController;
-import fi.otavanopisto.muikku.users.UserGroupEntityController;
 
 @Startup
 @Singleton
@@ -74,12 +68,6 @@ public class NeverLoggedInNotificationStrategy extends AbstractTimedNotification
   @Inject
   private UserController userController;
   
-  @Inject 
-  private UserGroupController userGroupController;
-  
-  @Inject
-  private UserGroupEntityController userGroupEntityController;
-  
   @Override
   public boolean isActive(){
     return active;
@@ -98,16 +86,15 @@ public class NeverLoggedInNotificationStrategy extends AbstractTimedNotification
       if (studentEntity != null) {
         Locale studentLocale = localeController.resolveLocale(LocaleUtils.toLocale(studentEntity.getLocale()));
         User student = userController.findUserByIdentifier(studentIdentifier);
-        User counselor = null;
-        UserEntity counselorEntity = getStudyCounselor(studentEntity.defaultSchoolDataIdentifier());
-        
-        if (counselorEntity != null) {
-        counselor = userController.findUserByIdentifier(counselorEntity.defaultSchoolDataIdentifier());
+        String guidanceCounselorEmail = notificationController.getStudyCounselorEmail(studentEntity.defaultSchoolDataIdentifier());
+        String notificationContent = localeController.getText(studentLocale, "plugin.timednotifications.notification.neverloggedin.content.defaultContent", new Object[] {student.getDisplayName()});
+        if (guidanceCounselorEmail != null) {
+        notificationContent = localeController.getText(studentLocale, "plugin.timednotifications.notification.neverloggedin.content.guidanceCounselorContent", new Object[] {student.getDisplayName(), guidanceCounselorEmail});
         }
         notificationController.sendNotification(
           localeController.getText(studentLocale, "plugin.timednotifications.notification.category"),
           localeController.getText(studentLocale, "plugin.timednotifications.notification.neverloggedin.subject"),
-          localeController.getText(studentLocale, "plugin.timednotifications.notification.neverloggedin.content", new Object[] {student.getDisplayName(), counselor.getDisplayName()}),
+          notificationContent,
           studentEntity,
           studentIdentifier,
           "neverloggedin"
@@ -119,43 +106,6 @@ public class NeverLoggedInNotificationStrategy extends AbstractTimedNotification
       }
       
     }
-  }
-  
-  private UserEntity getStudyCounselor(SchoolDataIdentifier studentIdentifier){
-    UserEntity guidanceCounselor = null;
-    List<UserGroupEntity> userGroupEntities = userGroupEntityController.listUserGroupsByUserIdentifier(studentIdentifier);
-    
-    // #3089: An awkward workaround to use the latest guidance group based on its identifier. Assumes a larger
-    // identifier means a more recent entity. A more proper fix would be to sync group creation dates from
-    // Pyramus and include them in the Elastic index. Then again, user groups would have to be refactored
-    // entirely, as Pyramus handles group members as students (one study programme) while Muikku handles
-    // them as user entities (all study programmes)...
-    
-    userGroupEntities.sort(new Comparator<UserGroupEntity>() {
-      public int compare(UserGroupEntity o1, UserGroupEntity o2) {
-        long l1 = NumberUtils.toLong(StringUtils.substringAfterLast(o1.getIdentifier(), "-"), -1);
-        long l2 = NumberUtils.toLong(StringUtils.substringAfterLast(o2.getIdentifier(), "-"), -1);
-        return (int) (l2 - l1);
-      }
-    });
-    
-    // Choose the first staff member of the first guidance group as CC recipient for the notification
-
-    userGroupEntities:
-    for (UserGroupEntity userGroupEntity : userGroupEntities) {
-      UserGroup userGroup = userGroupController.findUserGroup(userGroupEntity);
-
-      if (userGroup.getIsGuidanceGroup()) {
-        List<GroupUser> groupUsers = userGroupController.listUserGroupStaffMembers(userGroup);
-
-        for (GroupUser groupUser : groupUsers) {
-          User user = userGroupController.findUserByGroupUser(groupUser);
-          guidanceCounselor = userEntityController.findUserEntityByUser(user);
-          break userGroupEntities;
-        }
-      }
-    }
-    return guidanceCounselor;
   }
   
   private Collection<Long> getGroups(){
@@ -213,7 +163,7 @@ public class NeverLoggedInNotificationStrategy extends AbstractTimedNotification
 
       User student = userController.findUserByIdentifier(studentIdentifier);
       UserEntity studentEntity = userEntityController.findUserEntityByUser(student);
-      if ((student != null) && isNotifiedStudent(student.getStudyStartDate(), student.getStudyEndDate(), OffsetDateTime.now(), NOTIFICATION_THRESHOLD_DAYS)) {
+      if ((student != null) && notificationController.isNotifiedStudent(student.getStudyStartDate(), student.getStudyEndDate(), OffsetDateTime.now(), NOTIFICATION_THRESHOLD_DAYS)) {
 
         if (studentEntity.getLastLogin() == null && student.getStudyStartDate().isAfter(sendNotificationIfNotLoggedInBefore) && student.getStudyProgrammeName().equals("Nettilukio") || student.getStudyProgrammeName().equals("Nettiperuskoulu")) {
           studentIdentifiers.add(studentIdentifier);
@@ -222,20 +172,6 @@ public class NeverLoggedInNotificationStrategy extends AbstractTimedNotification
     }
     
     return studentIdentifiers;
-  }
-  
-  public boolean isNotifiedStudent(OffsetDateTime studyStartDate, OffsetDateTime studyEndDate, OffsetDateTime currentDateTime, int thresholdDays) {
-    if ((studyStartDate == null) || (studyEndDate != null))
-      return false;
-    
-    // Earliest point when student may receive the notification is study start date + threshold day count
-    OffsetDateTime thresholdDateTime = studyStartDate.plusDays(thresholdDays);
-    
-    // Furthest point to receive the notification is studyStartDate + thresholdDays + 30 (1 month)
-    OffsetDateTime maxThresholdDateTime = studyStartDate.plusDays(thresholdDays + 30);
-
-    // If the threshold date has passed the student is valid target for the notification
-    return thresholdDateTime.isBefore(currentDateTime) && maxThresholdDateTime.isAfter(currentDateTime);
   }
 
   private int offset = 0;
